% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{as.df}
\alias{as.df}
\title{Coerce to data frame}
\usage{
as.df(x)
}
\arguments{
\item{x}{Object to coerce with \code{as.data.frame()}.}
}
\value{
The coerced data frame.
}
\description{
Coerce to data frame
}
