% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{In}
\alias{In}
\title{Subset x to elements also present in y}
\usage{
In(x, y)
}
\arguments{
\item{x}{Vector to subset.}

\item{y}{Vector of values to keep.}
}
\value{
Elements of \code{x} that are also in \code{y}.
}
\description{
Subset x to elements also present in y
}
