% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{mprev}
\alias{mprev}
\title{Preview the top-left corner of a matrix}
\usage{
mprev(matrix, r = 2, c = 2)
}
\arguments{
\item{matrix}{Matrix to preview.}

\item{r, c}{Number of rows/columns to show.}
}
\value{
The \code{matrix[1:r, 1:c]} subset.
}
\description{
Preview the top-left corner of a matrix
}
