% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{NotIn}
\alias{NotIn}
\title{Subset x to elements NOT present in y}
\usage{
NotIn(x, y)
}
\arguments{
\item{x}{Vector to subset.}

\item{y}{Vector of values to exclude.}
}
\value{
Elements of \code{x} that are not in \code{y}.
}
\description{
Subset x to elements NOT present in y
}
