% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_list}
\alias{substitute_list}
\title{Substitute values in a config list}
\usage{
substitute_list(x)
}
\arguments{
\item{x}{Config list.}
}
\value{
The config list with values substituted
}
\description{
Substitute values in a config list
}
