% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_list}
\alias{substitute_list}
\title{Substitute values in a config list}
\usage{
substitute_list(x)
}
\arguments{
\item{x}{Config list.}
}
\value{
The config list with values substituted
}
\description{
Environment variables are already there (from outside and .Renviron),
\code{!expr} expressions and other R session level variables need to be present
at the time of config evaluation.
At last, the config level variables are evaluated, thus config level values can
refer to existing keys that are already substituted (i.e. not substituted from other
config values).
}
